# Performance enhancements for #141