--- a/valence_protocol/Cargo.toml
+++ b/valence_protocol/Cargo.toml
@@ -22,6 +22,11 @@ valence_nbt = { version = "0.5.0", path = "../valence_nbt" }
 
 [dev-dependencies]
 rand = "0.8.5"
+criterion = "0.4.0"
+
+[[bench]]
+name = "conversions"
+harness = false
 
 [build-dependencies]
 anyhow = "1.0.65"