--- a/valence_protocol/Cargo.toml
+++ b/valence_protocol/Cargo.toml
@@ -25,7 +25,7 @@ rand = "0.8.5"
 criterion = "0.4.0"
 
 [[bench]]
-name = "conversions"
+name = "benches"
 harness = false
 
 [build-dependencies]