@rj00a 2022-12-19
Nice work. I've added benchmarks for most of the other block functions so we can close both #141 and #74 now. 🎉